+++ src/core/index.ts
@@ -61,6 +61,7 @@ export {
   hasErrors,
   hasWarnings,
   map,
+  mapResult,
   mergeDiagnostics,
   mergeErrors,
   mergeWarnings,